chore: address issue #394
