+++ app/services/refraction_static_apply_trace_store.py
@@ -29,7 +29,11 @@ from app.services.refraction_static_artifacts import (
     REFRACTION_STATIC_QC_JSON_NAME,
     REFRACTION_STATIC_SOLUTION_NPZ_NAME,
 )
-from app.services.refraction_static_datum import RefractionDatumStaticsResult
+from app.services.refraction_static_types import (
+    RefractionDatumStaticsResult,
+    RefractionStaticApplyTraceStoreResult,
+    RefractionTraceShiftValidationResult,
+)
 from app.services.trace_store_index_validation import validate_sorted_to_original
 from app.services.trace_store_registration import (
     register_trace_store,
@@ -63,55 +67,6 @@ _OPTIONAL_SOLUTION_METADATA_KEYS = (
 
 class RefractionStaticTraceStoreApplyError(ValueError):
     """Raised when refraction statics cannot be applied to a TraceStore."""
-
-
-@dataclass(frozen=True)
-class RefractionTraceShiftValidationResult:
-    trace_shift_s_sorted: np.ndarray
-    trace_static_valid_mask_sorted: np.ndarray
-    trace_static_status_sorted: np.ndarray
-    trace_static_status_counts: dict[str, int]
-    max_abs_shift_ms: float
-    max_abs_applied_shift_ms: float
-    exceeds_max_abs_shift_count: int
-    n_valid_trace_shifts: int
-    n_invalid_trace_shifts: int
-    n_zero_trace_shifts: int
-    n_positive_trace_shifts: int
-    n_negative_trace_shifts: int
-
-
-@dataclass(frozen=True)
-class RefractionStaticApplyTraceStoreResult:
-    source_file_id: str
-    corrected_file_id: str | None
-
-    source_trace_store_path: Path
-    corrected_trace_store_path: Path | None
-
-    n_traces: int
-    n_samples: int
-    sample_interval_s: float
-
-    interpolation: str
-    fill_value: float
-    output_dtype: str
-
-    applied_shift_s_sorted: np.ndarray
-    applied_shift_ms_sorted: np.ndarray
-    trace_static_valid_mask_sorted: np.ndarray
-    trace_static_status_sorted: np.ndarray
-
-    max_abs_applied_shift_ms: float
-    n_valid_trace_shifts: int
-    n_invalid_trace_shifts: int
-    n_zero_trace_shifts: int
-    n_positive_trace_shifts: int
-    n_negative_trace_shifts: int
-
-    corrected_file_json: Path | None
-    qc_json: Path | None
-    qc: dict[str, Any]
 
 
 @dataclass(frozen=True)
