+++ app/services/refraction_static_weathering.py
@@ -14,8 +14,11 @@ import numpy as np
 from app.api.schemas import RefractionStaticApplyRequest, RefractionStaticModelRequest
 from app.core.state import AppState
 from app.services.refraction_static_half_intercept import (
-    RefractionHalfInterceptTimeResult,
     estimate_refraction_half_intercept_times_from_first_breaks,
+)
+from app.services.refraction_static_types import (
+    RefractionHalfInterceptTimeResult,
+    RefractionWeatheringThicknessResult,
 )
 
 REFRACTION_WEATHERING_QC_JSON_NAME = 'refraction_weathering_thickness_qc.json'
@@ -92,91 +95,6 @@ _TRACE_PREVIEW_COLUMNS = (
 
 class RefractionWeatheringThicknessError(ValueError):
     """Raised when weathering-thickness outputs cannot be built."""
-
-
-@dataclass(frozen=True)
-class RefractionWeatheringThicknessResult:
-    """Node, endpoint, trace-order, and QC output for GLI weathering thickness."""
-
-    bedrock_velocity_mode: Literal['solve_global', 'fixed_global']
-    bedrock_slowness_s_per_m: float
-    bedrock_velocity_m_s: float
-    weathering_velocity_m_s: float
-
-    node_id: np.ndarray
-    node_x_m: np.ndarray
-    node_y_m: np.ndarray
-    node_surface_elevation_m: np.ndarray
-    node_kind: np.ndarray
-
-    node_half_intercept_time_s: np.ndarray
-    node_half_intercept_time_ms: np.ndarray
-    node_weathering_thickness_m: np.ndarray
-    node_refractor_elevation_m: np.ndarray
-    node_solution_status: np.ndarray
-    node_weathering_status: np.ndarray
-
-    node_pick_count: np.ndarray
-    node_used_pick_count: np.ndarray
-    node_rejected_pick_count: np.ndarray
-    node_residual_rms_s: np.ndarray
-    node_residual_mad_s: np.ndarray
-
-    source_endpoint_key: np.ndarray
-    source_id: np.ndarray
-    source_node_id: np.ndarray
-    source_x_m: np.ndarray
-    source_y_m: np.ndarray
-    source_surface_elevation_m: np.ndarray
-    source_half_intercept_time_s: np.ndarray
-    source_weathering_thickness_m: np.ndarray
-    source_refractor_elevation_m: np.ndarray
-    source_weathering_status: np.ndarray
-
-    receiver_endpoint_key: np.ndarray
-    receiver_id: np.ndarray
-    receiver_node_id: np.ndarray
-    receiver_x_m: np.ndarray
-    receiver_y_m: np.ndarray
-    receiver_surface_elevation_m: np.ndarray
-    receiver_half_intercept_time_s: np.ndarray
-    receiver_weathering_thickness_m: np.ndarray
-    receiver_refractor_elevation_m: np.ndarray
-    receiver_weathering_status: np.ndarray
-
-    sorted_trace_index: np.ndarray
-    valid_observation_mask_sorted: np.ndarray
-    used_observation_mask_sorted: np.ndarray
-
-    source_endpoint_key_sorted: np.ndarray
-    receiver_endpoint_key_sorted: np.ndarray
-    source_node_id_sorted: np.ndarray
-    receiver_node_id_sorted: np.ndarray
-
-    source_half_intercept_time_s_sorted: np.ndarray
-    receiver_half_intercept_time_s_sorted: np.ndarray
-
-    source_weathering_thickness_m_sorted: np.ndarray
-    receiver_weathering_thickness_m_sorted: np.ndarray
-    source_refractor_elevation_m_sorted: np.ndarray
-    receiver_refractor_elevation_m_sorted: np.ndarray
-    source_weathering_status_sorted: np.ndarray
-    receiver_weathering_status_sorted: np.ndarray
-
-    estimated_first_break_time_s_sorted: np.ndarray
-    first_break_residual_s_sorted: np.ndarray
-
-    row_trace_index_sorted: np.ndarray
-    row_source_node_id: np.ndarray
-    row_receiver_node_id: np.ndarray
-    row_distance_m: np.ndarray
-    observed_pick_time_s: np.ndarray
-    modeled_pick_time_s: np.ndarray
-    residual_time_s: np.ndarray
-    used_row_mask: np.ndarray
-    rejected_by_robust_mask: np.ndarray
-
-    qc: dict[str, Any]
 
 
 @dataclass(frozen=True)
