+++ app/tests/test_refraction_static_datum.py
@@ -26,7 +26,7 @@ from app.services.refraction_static_datum import (
     compute_flat_datum_shift_s,
     compute_floating_datum_elevation_shift_s,
 )
-from app.services.refraction_static_weathering_replacement import (
+from app.services.refraction_static_types import (
     RefractionWeatheringReplacementStaticsResult,
 )
